server/requirements.txt
flask>=3.0.0
APScheduler>=3.10.4
waitress>=3.0.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
pyotp>=2.9.0
segno>=1.6.0
oracledb>=2.0.0
cryptography>=42.0.0
